requirements.txt
Django==4.2.20
django-cors-headers
python-dotenv
clerk-sdk-python